• Separates table header
